Moved macsafequeue definition and choice into parsl.multiprocessing module
